Разбивка на модули и правка по комментариям
